requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
groq==0.11.0
firebase-admin==6.5.0
python-multipart==0.0.9
pydantic==2.9.2
httpx==0.27.2
slowapi==0.1.9
websockets==13.1
google-genai>=1.0.0